requirements.txt
asgiref==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.6
django-crum==0.7.9
django-filter==2.4.0
django-import-export==2.5.0
django-qr-code==2.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
et-xmlfile==1.0.1
gunicorn==20.0.4
importlib-metadata==3.4.0
jdcal==1.4.1
Markdown==3.3.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
segno==1.3.1
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.7.4.3
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.0
